refactor: Extract the app state logic into lib
Don't require all ENV vars in debug mode
